+++ CHANGELOG.rst
@@ -1,5 +1,8 @@
 Ver. 0.9.2 - Development
 ===========================
+* Bug fix in :class:`.Links`
+* A more flexible implementation of the :meth:`~.Router.make_router` method
+  for the wsgi router
 
 Ver. 0.9.1 - 2014-Oct-14
 ===========================